Move test backend outside of testing package
